# Conflicts:
#	.idea/dataSources.local.xml
#	build/classes/java/main/it/unife/ingsw202324/EventManager/EventManagerApplication.class
#	build/classes/java/main/it/unife/ingsw202324/EventManager/api/CategoriesController.class
#	build/classes/java/main/it/unife/ingsw202324/EventManager/api/EventsController.class
#	build/classes/java/main/it/unife/ingsw202324/EventManager/models/Categories.class
#	build/classes/java/main/it/unife/ingsw202324/EventManager/models/Events.class
#	build/classes/java/main/it/unife/ingsw202324/EventManager/models/Organizers.class
#	build/classes/java/main/it/unife/ingsw202324/EventManager/models/TicketTypes.class
#	build/classes/java/main/it/unife/ingsw202324/EventManager/repositories/CategoriesRepository.class
#	build/classes/java/main/it/unife/ingsw202324/EventManager/repositories/TicketRepository.class
#	build/classes/java/main/it/unife/ingsw202324/EventManager/services/CategoriesService.class
#	build/classes/java/main/it/unife/ingsw202324/EventManager/services/EventService.class
#	build/classes/java/main/it/unife/ingsw202324/EventManager/services/TemplateRestConsumer.class
#	build/libs/MicroservizioBase-0.0.1-SNAPSHOT-plain.jar
#	build/libs/MicroservizioBase-0.0.1-SNAPSHOT.jar
#	build/reports/tests/test/classes/it.unife.ingsw202324.EventManager.EventManagerApplicationTests.html
#	build/reports/tests/test/index.html
#	build/reports/tests/test/packages/it.unife.ingsw202324.EventManager.html
#	build/test-results/test/TEST-it.unife.ingsw202324.EventManager.EventManagerApplicationTests.xml
#	build/test-results/test/binary/output.bin
#	build/test-results/test/binary/output.bin.idx
#	build/test-results/test/binary/results.bin
#	build/tmp/compileTestJava/previous-compilation-data.bin

+++ src/main/java/it/unife/ingsw202324/EventManager/api/CategoriesController.java
@@ -42,7 +42,7 @@ public class CategoriesController {
 
     // Restituisce una categoria dato il suo id
     @GetMapping("/{id}")
-    public ResponseEntity<Categories> getCategoryById(@PathVariable Long id) {
+    public ResponseEntity<Categories> getCategoryById(@PathVariable int id) {
         Optional<Categories> category = categoriesService.getCategoryById(id);
         if (category.isPresent()) {
             return ResponseEntity.ok(category.get());
@@ -57,7 +57,7 @@ public class CategoriesController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Categories> updateCategory(@PathVariable Long id, @RequestBody Categories categoryDetails) {
+    public ResponseEntity<Categories> updateCategory(@PathVariable int id, @RequestBody Categories categoryDetails) {
         Categories updatedCategory = categoriesService.updateCategory(id, categoryDetails);
         if (updatedCategory != null) {
             return ResponseEntity.ok(updatedCategory);
@@ -68,7 +68,7 @@ public class CategoriesController {
 
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCategory(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteCategory(@PathVariable int id) {
         categoriesService.deleteCategory(id);
         return ResponseEntity.noContent().build();
     }
